Further develop layout structure and settled on page format
